+++ extension-androidauto/src/main/java/com/mapbox/maps/extension/androidauto/MapboxCarMapInitializer.kt
@@ -1,0 +1,20 @@
+package com.mapbox.maps.extension.androidauto
+
+import androidx.car.app.CarContext
+import androidx.car.app.Session
+import androidx.lifecycle.Lifecycle
+import com.mapbox.maps.MapInitOptions
+
+/**
+ * Lazy initialization for the [MapInitOptions]. The [CarContext] is not created until the [Session]
+ * has been created. Use this initializer to specify custom initial map options.
+ */
+fun interface MapboxCarMapInitializer {
+  /**
+   * Triggered once the [Session] reaches the [Lifecycle.State.CREATED] state.
+   *
+   * @param carContext from [Session.getCarContext].
+   * @return [MapInitOptions] for the initial [MapboxCarMap] configuration.
+   */
+  fun onCreate(carContext: CarContext): MapInitOptions
+}