+++ extension-androidauto/src/test/java/com/mapbox/maps/extension/androidauto/testing/TestLifecycleOwner.kt
@@ -1,0 +1,16 @@
+package com.mapbox.maps.extension.androidauto.testing
+
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.LifecycleRegistry
+
+class TestLifecycleOwner : LifecycleOwner {
+  private val lifecycleRegistry = LifecycleRegistry(this)
+    .also { it.currentState = Lifecycle.State.INITIALIZED }
+
+  override fun getLifecycle(): Lifecycle = lifecycleRegistry
+
+  fun moveToState(state: Lifecycle.State) {
+    lifecycleRegistry.currentState = state
+  }
+}